Move events configuration to monolith
