+++ monolith/src/main/java/io/github/mat3e/fairytales/EventsConfiguration.java
@@ -1,13 +1,15 @@
-package io.github.mat3e.fairytales.pigs3.model;
+package io.github.mat3e.fairytales;
 
 import io.github.mat3e.ddd.event.DomainEvent;
 import io.github.mat3e.ddd.event.DomainEventPublisher;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableAsync;
 
 @Configuration
-class ModelConfiguration {
+class EventsConfiguration {
     @Bean
     DomainEventPublisher eventPublisher(final ApplicationEventPublisher applicationEventPublisher) {
         return new SpringDomainEventPublisher(applicationEventPublisher);
@@ -24,5 +26,11 @@ class ModelConfiguration {
         public void publish(final DomainEvent event) {
             applicationEventPublisher.publishEvent(event);
         }
+    }
+
+    @EnableAsync
+    @Configuration
+    @ConditionalOnProperty(value = "spring.main.web-application-type", havingValue = "servlet")
+    static class AsyncConfiguration {
     }
 }
