+++ pigs3/adapters/src/main/java/io/github/mat3e/fairytales/pigs3/app/AppConfiguration.java
@@ -3,12 +3,10 @@ package io.github.mat3e.fairytales.pigs3.app;
 import io.github.mat3e.ddd.event.DomainEventPublisher;
 import io.github.mat3e.fairytales.pigs3.model.HouseRepository;
 import io.github.mat3e.fairytales.pigs3.model.event.HouseEvent;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.event.EventListener;
 import org.springframework.scheduling.annotation.Async;
-import org.springframework.scheduling.annotation.EnableAsync;
 
 @Configuration(proxyBeanMethods = false)
 class AppConfiguration {
@@ -26,12 +24,7 @@ class AppConfiguration {
     @Async
     @EventListener
     public void on(final HouseEvent event) {
+        System.out.println(Thread.currentThread().getName());
         app.getEventHandler().handle(event);
     }
-}
-
-@EnableAsync
-@Configuration
-@ConditionalOnProperty(value = "spring.main.web-application-type", havingValue = "servlet")
-class AsyncConfiguration {
 }
