+++ pigs3/adapters/src/test/java/io/github/mat3e/fairytales/TestApp.java
@@ -1,7 +1,14 @@
 package io.github.mat3e.fairytales;
 
+import io.github.mat3e.ddd.event.DomainEventPublisher;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.context.annotation.Bean;
 
 @SpringBootApplication
 class TestApp {
+    @Bean
+    DomainEventPublisher testPublisher(ApplicationEventPublisher applicationEventPublisher) {
+        return applicationEventPublisher::publishEvent;
+    }
 }
